+++ src/main/java/racingcar/RacingCarApplication.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.RacingCarGame;
+
 public class RacingCarApplication {
 
     public static void main(String[] args) {
