+++ src/main/java/racingcar/view/Console.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
+import racingcar.common.RacingCarException;
 
 public class Console {
