+++ src/main/java/racingcar/view/PrintResult.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.RacingCar;
 
 public class PrintResult {
